fix: ensure float32 precision in embedding rotation calculations

+++ lib/python/openlinktoken-core-ai/src/main/openlinktoken_core_ai/tokentransformer/rotation/embedding_rotator.py
@@ -29,14 +29,20 @@ def rotate(
     Returns:
         List of k-dimensional float lists, one per entry in ``matrices``.
     """
-    x_centered = np.asarray(embedding, dtype=np.float64) - np.asarray(bias, dtype=np.float64)
+    # Cast the input to float32 first, then promote to float64 for the arithmetic.
+    # This matches the Java EmbeddingRotator contract (float[] embedding → double arithmetic)
+    # and ensures cross-language parity: both sides start from identical float32 precision
+    # regardless of whether the caller holds the embedding as float32 or float64.
+    x_centered = np.asarray(embedding, dtype=np.float32).astype(np.float64) - np.asarray(bias, dtype=np.float64)
     results = []
     for R in matrices:
         R_arr = R if isinstance(R, np.ndarray) else np.asarray(R, dtype=np.float64)
         if R_arr.size == 1 and R_arr.ravel()[0] == -1.0:
             # Sentinel: pass-through — return first k values of the centred embedding.
-            results.append(x_centered[:k].tolist())
+            # Cast back to float32 to match Java's (float) xCentered[i] cast.
+            results.append(x_centered[:k].astype(np.float32).tolist())
         else:
             # Q[:k, :] @ x  — only the first k rows of the rotation matrix are needed.
-            results.append((R_arr[:k, :] @ x_centered).tolist())
+            # Cast back to float32 to match Java's (float) sum cast.
+            results.append((R_arr[:k, :] @ x_centered).astype(np.float32).tolist())
     return results
